pyproject.toml
[project]
name = "toio-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "toio-py>=1.1.0",
]
